Reintroduced intermediate interfaces
