+++ documentation/pom.xml
@@ -117,7 +117,7 @@
 									<inline>
 										<files>
 											<file>
-												<source>${project.basedir}/src/main/resources/helloservice.yaml</source>
+												<source>${project.basedir}/src/main/resources/api.yaml</source>
 												<destName>${git.commit.id.describe}.yaml</destName>
 											</file>
 											<file>
@@ -160,7 +160,7 @@
 							<goal>generate</goal>
 						</goals>
 						<configuration>
-							<inputSpec>${project.basedir}/src/main/resources/helloservice.yaml</inputSpec>
+							<inputSpec>${project.basedir}/src/main/resources/api.yaml</inputSpec>
 							<generatorName>java</generatorName>
 							<configOptions>
 								<sourceFolder>src/test/java</sourceFolder>
@@ -179,7 +179,7 @@
 							<goal>generate</goal>
 						</goals>
 						<configuration>
-							<inputSpec>${project.basedir}/src/main/resources/helloservice.yaml</inputSpec>
+							<inputSpec>${project.basedir}/src/main/resources/api.yaml</inputSpec>
 							<generatorName>spring</generatorName>
 							<configOptions>
 								<sourceFolder>src/main/java</sourceFolder>
