+++ service/src/test/java/dk/kvalitetsit/hello/configuration/TestConfiguration.java
@@ -1,6 +1,6 @@
 package dk.kvalitetsit.hello.configuration;
 
-import dk.kvalitetsit.hello.dao.HelloDao;
+import dk.kvalitetsit.hello.dao.HelloDaoImpl;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -13,7 +13,7 @@ public class TestConfiguration {
     // Configure beans used for test
 
     @Bean
-    public HelloDao helloDao(DataSource dataSource) {
-        return new HelloDao(dataSource);
+    public HelloDaoImpl helloDao(DataSource dataSource) {
+        return new HelloDaoImpl(dataSource);
     }
 }
