+++ service/src/test/java/dk/kvalitetsit/hello/dao/HelloDaoImplTest.java
@@ -9,7 +9,7 @@ import java.util.UUID;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-public class HelloDaoTest extends AbstractDaoTest {
+public class HelloDaoImplTest extends AbstractDaoTest {
     @Autowired
     private HelloDao helloDao;
